Move TokenMapWithScripts to primitive
