+++ lib/primitive/cardano-wallet-primitive.cabal
@@ -56,6 +56,7 @@ library
     , cardano-ledger-mary
     , cardano-ledger-shelley
     , cardano-numeric
+    , cardano-protocol-tpraos
     , cardano-slotting
     , cardano-strict-containers
     , cardano-wallet-read
@@ -139,6 +140,7 @@ library
     Cardano.Wallet.Primitive.Types.TokenBundleMaxSize
     Cardano.Wallet.Primitive.Types.TokenMap
     Cardano.Wallet.Primitive.Types.TokenMap.Gen
+    Cardano.Wallet.Primitive.Types.TokenMapWithScripts
     Cardano.Wallet.Primitive.Types.TokenPolicy
     Cardano.Wallet.Primitive.Types.TokenPolicy.Gen
     Cardano.Wallet.Primitive.Types.TokenQuantity
