+++ lib/wallet/src/Cardano/Wallet/Transaction.hs
@@ -66,7 +66,6 @@ import Cardano.Address.Script
     ( KeyHash (..)
     , KeyRole (..)
     , Script (..)
-    , ScriptHash
     , ScriptTemplate
     )
 import Cardano.Api
@@ -107,8 +106,14 @@ import Cardano.Wallet.Primitive.Types.TokenMap
     ( AssetId
     , TokenMap
     )
-import Cardano.Wallet.Primitive.Types.TokenPolicy
-    ( TokenPolicyId
+import Cardano.Wallet.Primitive.Types.TokenMapWithScripts
+    ( AnyScript (..)
+    , PlutusScriptInfo (..)
+    , PlutusVersion (..)
+    , ReferenceInput (..)
+    , ScriptReference (..)
+    , TokenMapWithScripts (..)
+    , emptyTokenMapWithScripts
     )
 import Cardano.Wallet.Primitive.Types.TokenQuantity
     ( TokenQuantity
@@ -141,20 +146,11 @@ import Data.Map.Strict
 import Data.Monoid.Monus
     ( (<\>)
     )
-import Data.Quantity
-    ( Quantity (..)
-    )
 import Data.Text
     ( Text
     )
-import Data.Text.Class
-    ( FromText (..)
-    , TextDecodingError (..)
-    , ToText (..)
-    )
 import Data.Word
-    ( Word64
-    , Word8
+    ( Word8
     )
 import Fmt
     ( Buildable (..)
@@ -334,64 +330,6 @@ data DelegationAction
 
 instance Buildable DelegationAction where
     build = genericF
-
-data PlutusVersion = PlutusVersionV1 | PlutusVersionV2 | PlutusVersionV3
-    deriving (Eq, Generic, Show)
-    deriving anyclass NFData
-
-instance ToText PlutusVersion where
-    toText PlutusVersionV1 = "v1"
-    toText PlutusVersionV2 = "v2"
-    toText PlutusVersionV3 = "v3"
-
-instance FromText PlutusVersion where
-    fromText txt = case txt of
-        "v1" -> Right PlutusVersionV1
-        "v2" -> Right PlutusVersionV2
-        "v3" -> Right PlutusVersionV3
-        _ -> Left $ TextDecodingError $ unwords
-            [ "I couldn't parse the given plutus version."
-            , "I am expecting one of the words 'v1' or"
-            , "'v2'."]
-
-data PlutusScriptInfo = PlutusScriptInfo
-    { languageVersion :: PlutusVersion
-    , scriptHash :: ScriptHash
-    }
-    deriving (Eq, Generic, Show)
-    deriving anyclass NFData
-
-newtype ReferenceInput = ReferenceInput TxIn
-    deriving (Eq, Generic, Show)
-    deriving anyclass NFData
-
--- | ScriptReference depicts whether the script is referenced via spending
--- and is bound to be used in the same transaction or is referenced via
--- reference inputs and is to be used in other transactions. The the latter
--- case the script is referenced in other trasactions
-data ScriptReference =
-      ViaSpending
-    | ViaReferenceInput ReferenceInput
-    deriving (Eq, Generic, Show)
-    deriving anyclass NFData
-
-data AnyScript =
-      NativeScript !(Script KeyHash) !ScriptReference
-    | PlutusScript !PlutusScriptInfo !ScriptReference
-    | AnyScriptReference !ScriptHash ![ReferenceInput]
-    deriving (Eq, Generic, Show)
-    deriving anyclass NFData
-
-data TokenMapWithScripts = TokenMapWithScripts
-    { txTokenMap :: !TokenMap
-    , txScripts :: !(Map TokenPolicyId AnyScript)
-    } deriving (Show, Generic, Eq)
-
-emptyTokenMapWithScripts :: TokenMapWithScripts
-emptyTokenMapWithScripts = TokenMapWithScripts
-    { txTokenMap = mempty
-    , txScripts = Map.empty
-    }
 
 data AnyExplicitScript =
       NativeExplicitScript !(Script KeyHash) !ScriptReference
